+++ src/client/src/mui.theme.d.ts
@@ -137,6 +137,7 @@ declare module "@mui/material/styles" {
     MuiDialogContentText: object;
     MuiTableCell: object;
     MuiTooltip: object;
+    MuiDataGrid: object;
   }
 
   interface AppThemeComponentsOptions extends ComponentsOptions {
@@ -160,6 +161,7 @@ declare module "@mui/material/styles" {
     MuiDialogContentText: object;
     MuiTableCell: object;
     MuiTooltip: object;
+    MuiDataGrid: object;
   }
 
   interface AppTheme extends Theme {